Igualando branch

+++ database/migrations/2022_12_28_002551_create_table_google_credentials.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text('google_client_secret')->nullable();
             $table->text('google_redirect_uri')->nullable();
             $table->text('google_webhook_uri')->nullable();
-            $table->integer('user_id')->unsigned();;
+            $table->bigInteger('user_id')->unsigned();;
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
